got the kth nearest AI generally working. The way to record inputs needs to be improved

+++ ResearchGame/Assets/AIScripts/Timer.cs
@@ -8,6 +8,6 @@ public class Timer : MonoBehaviour {
 	
 	// Update is called once per frame
 	void Update () {
-        currentText.text = GameManager.timeRemaining.ToString();
+        currentText.text = ((int)(GameManager.timeRemaining)).ToString();
 	}
 }
